+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All important changes to the decitala package will be documented here.
 The changelog format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and the project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [v0.8.0] ??? 
+### Added
+- Added a ``loader`` function in utils.py for easy loading of the valid analyses. 
 
 ## [v0.7.4] March 8, 2021 (Kent)
 ### Fixed